3370 - Move MultiSelectOption component to its own file
