+++ src/client/components/Inputs/Select/Indicators/Indicators.tsx
@@ -1,19 +1,10 @@
 import './Indicators.scss'
-import React, { useEffect, useRef } from 'react'
-import {
-  ClearIndicatorProps,
-  components,
-  DropdownIndicatorProps,
-  IndicatorsContainerProps,
-  OptionProps,
-} from 'react-select'
+import React from 'react'
+import { ClearIndicatorProps, components, DropdownIndicatorProps, IndicatorsContainerProps } from 'react-select'
 
 import classNames from 'classnames'
 
 import Icon from 'client/components/Icon'
-import { Option } from 'client/components/Inputs/Select'
-
-import { useMultiSelectOptionConfig } from './hooks/useMultiSelectOptionConfig'
 
 export const IndicatorsContainer: React.FC<IndicatorsContainerProps> = (props) => {
   const { children, className, isDisabled, ...rest } = props
@@ -48,33 +39,5 @@ export const ClearIndicator: React.FC<ClearIndicatorProps> = (props) => {
     <components.ClearIndicator {...rest} className={classNames(className, 'select__clearIndicator')}>
       <Icon className="icon-sub" name="remove" />
     </components.ClearIndicator>
-  )
-}
-
-export const MultiSelectOption: React.FC<OptionProps<Option>> = (props) => {
-  const { data, label } = props
-
-  const { checked, isInputIndeterminate } = useMultiSelectOptionConfig(props)
-
-  const inputRef = useRef<HTMLInputElement>(null)
-  useEffect(() => {
-    if (inputRef.current) {
-      inputRef.current.indeterminate = isInputIndeterminate
-    }
-  }, [isInputIndeterminate])
-
-  return (
-    // eslint-disable-next-line react/jsx-props-no-spreading
-    <components.Option {...props}>
-      <input
-        key={`${data.value}-${isInputIndeterminate}`}
-        ref={inputRef}
-        checked={checked}
-        className="select__toggleAllOption-checkbox"
-        onChange={() => undefined}
-        type="checkbox"
-      />
-      <span className="select__toggleAllOption-label">{label}</span>
-    </components.Option>
   )
 }
