+++ src/client/components/Inputs/Select/Indicators/MultiSelectOption/index.ts
@@ -1,0 +1,1 @@
+export { MultiSelectOption } from './MultiSelectOption'
