+++ src/client/components/Inputs/Select/Indicators/index.ts
@@ -1,1 +1,2 @@
-export { ClearIndicator, DropdownIndicator, IndicatorsContainer, MultiSelectOption } from './Indicators'
+export { ClearIndicator, DropdownIndicator, IndicatorsContainer } from './Indicators'
+export { MultiSelectOption } from './MultiSelectOption'
